Exercice 2No bogue
It has no bogue. Just Netbeans don't accpt "2.5" flot but "2,5" float in FR computer.

+++ ExercicesPOO/src/main/java/ch/conceptforge/exercicespoo/Exercice02/NumberBot.java
@@ -29,7 +29,7 @@ public class NumberBot {
         System.out.println(askReal);
         unReal = sc.nextFloat();
         
-        maSomme = unEntier + unReal;
+        maSomme = (float) unEntier + unReal;
         maMoy = maSomme / 2;
